subiendo cambios de compras y ventas modificado

+++ WebVentas/src/main/java/com/webventas/domain/dto/request/CompraTransaccionRequest.java
@@ -2,13 +2,14 @@ package com.webventas.domain.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class CompraTransaccionRequest {
     private Long idProducto;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double precioVenta;
-    private Double importe;
+    private BigDecimal precioUnitario;
+    private BigDecimal precioVenta;
+    private BigDecimal importe;
 }
